pyproject.toml
[tool.poetry]
name = "mizuscripts"
version = "0.1.0"
description = ""
authors = ["shudong <dongs2011@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
scikit-learn = "^1.5.0"
chromadb = "^0.5.3"
openai = "^1.35.3"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
mongoengine = "^0.28.2"
boto3 = "^1.34.131"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
coverage = "^7.5.3"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.26.1"

[tool.poe.tasks]
cli.script = "mizuscripts.cli:main" # python script based task

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
